requirements.txt
scratchattach
groq
python-dotenv